requirements.txt
sentence-transformers
faiss-cpu
numpy
pandas
PyPDF2
tqdm
streamlit
langchain
transformers
torch
pymupdf
openai
python-dotenv
streamlit
rank-bm25
networkx
